debt: Extracted filelist components from FolderView
