--- src/drive/styles/folderview.styl
@@ -1,1 +1,0 @@
-@import './filelist'
