+++ src/drive/web/modules/filelist/AsyncFileListRows.jsx
@@ -2,8 +2,8 @@ import React from 'react'
 import PropTypes from 'prop-types'
 import Oops from 'components/Error/Oops'
 import { EmptyDrive, EmptyTrash } from 'components/Error/Empty'
-import FileListPlaceholder from './FileListPlaceholder'
-import FileList from './FileList'
+import FileListRowsPlaceholder from './FileListRowsPlaceholder'
+import FileListRows from './FileListRows'
 
 const EmptyContent = props => {
   const { isTrashContext, canUpload } = props
@@ -25,28 +25,28 @@ EmptyContent.defaultProps = {
   params: {}
 }
 
-const FolderContent = props => {
+const AsyncFileListRows = props => {
   const { files, isAddingFolder, isLoading, isInError } = props
 
-  if (isLoading) return <FileListPlaceholder />
+  if (isLoading) return <FileListRowsPlaceholder />
   else if (isInError) return <Oops />
   else if (files.length === 0 && !isAddingFolder)
     return <EmptyContent {...props} />
-  else return <FileList withSelectionCheckbox {...props} />
+  else return <FileListRows withSelectionCheckbox {...props} />
 }
 
-FolderContent.propTypes = {
+AsyncFileListRows.propTypes = {
   files: PropTypes.array,
   isAddingFolder: PropTypes.bool,
   isLoading: PropTypes.bool,
   isInError: PropTypes.bool
 }
 
-FolderContent.defaultProps = {
+AsyncFileListRows.defaultProps = {
   files: [],
   isAddingFolder: false,
   isLoading: false,
   isInError: false
 }
 
-export default FolderContent
+export default AsyncFileListRows
