+++ src/drive/web/modules/layout/FolderView.jsx
@@ -1,7 +1,6 @@
 /* global __TARGET__ */
 import React, { Component } from 'react'
 import { translate } from 'cozy-ui/react/I18n'
-import Alerter from 'cozy-ui/react/Alerter'
 import { ModalManager } from 'react-cozy-helpers'
 
 import { ROOT_DIR_ID } from 'drive/constants/config'
@@ -9,12 +8,9 @@ import { ROOT_DIR_ID } from 'drive/constants/config'
 import MediaBackupProgression from 'drive/mobile/modules/mediaBackup/MediaBackupProgression'
 import RatingModal from 'drive/mobile/modules/settings/RatingModal'
 import FirstUploadModal from 'drive/mobile/modules/mediaBackup/FirstUploadModal'
-import FileListHeader, {
-  MobileFileListHeader
-} from 'drive/web/modules/filelist/FileListHeader'
+
 import FileActionMenu from 'drive/web/modules/actionmenu/FileActionMenu'
-import FolderContent from 'drive/web/modules/filelist/FolderContent'
-import AddFolder from 'drive/web/modules/filelist/AddFolder'
+import FileList from 'drive/web/modules/filelist/FileList'
 import Breadcrumb from 'drive/web/modules/navigation/Breadcrumb'
 import SelectionBar from 'drive/web/modules/selection/SelectionBar'
 import Dropzone from 'drive/web/modules/upload/Dropzone'
@@ -22,32 +18,7 @@ import { getFolderIdFromRoute } from 'drive/web/modules/navigation/duck'
 import Main from './Main'
 import Topbar from './Topbar'
 
-import styles from 'drive/styles/folderview'
-
-const toggle = (flag, state, props) => ({ [flag]: !state[flag] })
-
 class FolderView extends Component {
-  state = {
-    showAddFolder: false
-  }
-
-  toggleAddFolder = () => {
-    this.setState(toggle.bind(null, 'showAddFolder'))
-  }
-
-  createFolder = name => {
-    return this.props.actions.list
-      .createFolder(name)
-      .then(() => this.toggleAddFolder())
-  }
-
-  abortAddFolder = accidental => {
-    if (accidental) {
-      Alerter.info('alert.folder_abort')
-    }
-    this.toggleAddFolder()
-  }
-
   render() {
     const {
       children,
@@ -68,8 +39,6 @@ class FolderView extends Component {
       canCreateFolder
     } = this.props
     const { hideActionMenu, showSelectionBar, uploadFiles } = this.props
-
-    const { showAddFolder } = this.state
 
     const fetchFailed = this.props.fetchStatus === 'failed'
     const fetchPending = this.props.fetchStatus === 'pending'
@@ -113,26 +82,13 @@ class FolderView extends Component {
           <div style={{ display: selectionModeActive ? 'inherit' : 'none' }}>
             <SelectionBar selected={selected} actions={actions.selection} />
           </div>
-
-          <div className={styles['fil-content-table']} role="table">
-            <MobileFileListHeader canSort={canSort} />
-            <FileListHeader canSort={canSort} />
-            <div className={styles['fil-content-body']}>
-              {showAddFolder && (
-                <AddFolder
-                  onSubmit={this.createFolder}
-                  onAbort={this.abortAddFolder}
-                />
-              )}
-              <FolderContent
-                {...this.props}
-                selectionModeActive={selectionModeActive}
-                isAddingFolder={showAddFolder}
-                isLoading={fetchPending || isNavigating}
-                isInError={fetchFailed}
-              />
-            </div>
-          </div>
+          <FileList
+            {...this.props}
+            canSort={canSort}
+            selectionModeActive={selectionModeActive}
+            isLoading={fetchPending || isNavigating}
+            isInError={fetchFailed}
+          />
           {this.renderViewer(children)}
           {actionMenuActive && (
             <FileActionMenu
